requirements.txt
Flask==2.2.2
google-cloud-datastore==2.10.0
python-jose
authlib
requests==2.28.1
python-dotenv==0.21.0